Check projection

+++ prep_rasters.py
@@ -53,36 +53,42 @@ def raster_bands(tif, sub):
     error_email('Cannot read ' + tif_file)
   
   else:
-    if (src.RasterCount == 4 or (val != 0 and val != 255)) and ras.dtypes[0] == 'uint8':
-      print('4 correct bands found')
-      logger.info('4 correct bands found')
-      copyfile(tif_file, PATH + 'converted/' + tif)
-      project_raster(tif)
-    elif src.RasterCount == 1 or src.RasterCount == 3 or src.RasterCount == 4:
-      gdal_string = Template("""gdal_translate -b 1 ${f} -a_nodata none ${path}converted/red.tif &&
-          gdal_translate -b ${b2} ${f} -a_nodata none ${path}converted/green.tif &&
-          gdal_translate -b ${b3} ${f} -a_nodata none ${path}converted/blue.tif &&
-          echo Calculating file mask
-          gdal_calc.py -A ${path}converted/red.tif -B ${path}converted/green.tif -C ${path}converted/blue.tif --outfile=${path}converted/mask.tif --calc="logical_and(A!=${nodata},B!=${nodata},C!=${nodata})*255" --NoDataValue=0 &&
-          echo Merging files
-          gdal_merge.py -separate -ot Byte -o ${path}converted/${tif} ${path}converted/red.tif ${path}converted/green.tif ${path}converted/blue.tif ${path}converted/mask.tif &&
-          echo Cleaning up
-          rm ${path}converted/red.tif &&
-          rm ${path}converted/green.tif &&
-          rm ${path}converted/blue.tif &&
-          rm ${path}converted/mask.tif""")
-      if src.RasterCount == 1:
-        print('1 band raster found')
-        logger.info('1 band raster found')
-        os.system(gdal_string.substitute(f=tif_file, b2='1', b3='1', nodata=str(val),
-                                          path=PATH, tif=tif))
+    epsg = int(gdal.Info(src, format='json')['coordinateSystem']['wkt'].rsplit('"EPSG","', 1)[-1].split('"')[0])
+    if epsg == 4326:
+      if (src.RasterCount == 4 or (val != 0 and val != 255)) and ras.dtypes[0] == 'uint8':
+        print('4 correct bands found')
+        logger.info('4 correct bands found')
+        copyfile(tif_file, PATH + 'converted/' + tif)
         project_raster(tif)
-      elif src.RasterCount >= 3:
-        print('3+ band raster found')
-        logger.info('3+ band raster found')
-        os.system(gdal_string.substitute(f=tif_file, b2='2', b3='3', nodata=str(val),
-                                          path=PATH, tif=tif))
-        project_raster(tif)
+      elif src.RasterCount == 1 or src.RasterCount == 3 or src.RasterCount == 4:
+        gdal_string = Template("""gdal_translate -b 1 ${f} -a_nodata none ${path}converted/red.tif &&
+            gdal_translate -b ${b2} ${f} -a_nodata none ${path}converted/green.tif &&
+            gdal_translate -b ${b3} ${f} -a_nodata none ${path}converted/blue.tif &&
+            echo Calculating file mask
+            gdal_calc.py -A ${path}converted/red.tif -B ${path}converted/green.tif -C ${path}converted/blue.tif --outfile=${path}converted/mask.tif --calc="logical_and(A!=${nodata},B!=${nodata},C!=${nodata})*255" --NoDataValue=0 &&
+            echo Merging files
+            gdal_merge.py -separate -ot Byte -o ${path}converted/${tif} ${path}converted/red.tif ${path}converted/green.tif ${path}converted/blue.tif ${path}converted/mask.tif &&
+            echo Cleaning up
+            rm ${path}converted/red.tif &&
+            rm ${path}converted/green.tif &&
+            rm ${path}converted/blue.tif &&
+            rm ${path}converted/mask.tif""")
+        if src.RasterCount == 1:
+          print('1 band raster found')
+          logger.info('1 band raster found')
+          os.system(gdal_string.substitute(f=tif_file, b2='1', b3='1', nodata=str(val),
+                                            path=PATH, tif=tif))
+          project_raster(tif)
+        elif src.RasterCount >= 3:
+          print('3+ band raster found')
+          logger.info('3+ band raster found')
+          os.system(gdal_string.substitute(f=tif_file, b2='2', b3='3', nodata=str(val),
+                                            path=PATH, tif=tif))
+          project_raster(tif)
+      else:
+        print(tif_file + ' is in the wrong projection!')
+        logger.warning(tif_file + ' is in the wrong projection', level="WARN")
+        error_email(tif_file + ' is in the wrong projection')
     else:
       print(tif_file + ' has wrong number of bands!')
       logger.warning(tif_file + ' has wrong number of bands', level="WARN")
